refactor: input and customItem have own component
